Make amd / umd configurable. Closes #262.
