+++ ecmascript/transforms/src/modules/mod.rs
@@ -1,5 +1,6 @@
+#[macro_use]
+mod util;
 pub mod amd;
 pub mod common_js;
 pub mod import_analysis;
 pub mod umd;
-mod util;
